蛋e网保存csv msyql

+++ Egg_e_web/merge_data.py
@@ -5,7 +5,6 @@ savePath = 'D://CSDY_data/蛋e网各省价格一览'
 filePath = 'D://CSDY_data/蛋e网蛋价'
 header = ('省份','城市', '蛋种', '销售方式', '价格/元', '单位/斤', '趋势', '其他信息', '时间')
 provinces = os.listdir(filePath)
-import pandas as pd
 
 import pymysql
 
@@ -36,37 +35,51 @@ def data2mysql(data):
         print("插入成功")
     return (db,cursor)
 
-
-result=[]
-
-for province in provinces:
-    province_datas_path = filePath + "/" + province
-    data_list = os.listdir(province_datas_path)
-    data_list.sort(reverse=True)
-    with open(savePath + '/' + '%s.csv' % province, 'w+', newline='', encoding='UTF-8') as csvfile:
-        writer = csv.writer(csvfile)
-        writer.writerow(list(header))
-        for data in data_list:
-            # if (os.path.isdir(filePath + '/'+province+'/' + data)):  # 判断是否是文件夹
-            if (data[0] == '.'):  # 排除隐藏文件夹
-                pass
-            else:  # 添加非隐藏文件
-                date_time = data.split('.')[0]
-                df = open(filePath + '/' + province + '/' + data, encoding="GBK")
-                for raw in df:
-                    if "城市" in raw:
-                        pass
-                    else:
-                        raw = province+','+raw.replace('\n', ' ') + ',' + date_time
-                        split = list(raw.split(','))
-                        print(split)
-                        result.append(split)
-        writer.writerows(result)
-        print(province+" finished")
-        mysql = data2mysql(result)
-        mysql[0].commit()
-        mysql[0].close()
-        mysql[1].close()
-        result.clear()
+def save_csv(province,data):
+    if os.path.exists(savePath + '/' + province + '.csv'):
+        with open(savePath + '/' + '%s.csv' % province, 'a', newline='', encoding='UTF-8') as csvfile:
+            writer = csv.writer(csvfile)
+            writer.writerows(data)
+    else:
+        with open(savePath + '/' + '%s.csv' % province, 'w+', newline='', encoding='UTF-8') as csvfile:
+            writer = csv.writer(csvfile)
+            writer.writerow(list(header))
+            writer.writerows(data)
     csvfile.close()
-    print(province + " 完成")
+    print(savePath + '/' + province + '.csv'+'  写入完成')
+
+
+def save():
+    result=[]
+    for province in provinces:
+        province_datas_path = filePath + "/" + province
+        data_list = os.listdir(province_datas_path)
+        data_list.sort(reverse=True)
+        with open(savePath + '/' + '%s.csv' % province, 'w+', newline='', encoding='UTF-8') as csvfile:
+            writer = csv.writer(csvfile)
+            writer.writerow(list(header))
+            for data in data_list:
+                # if (os.path.isdir(filePath + '/'+province+'/' + data)):  # 判断是否是文件夹
+                if (data[0] == '.'):  # 排除隐藏文件夹
+                    pass
+                else:  # 添加非隐藏文件
+                    date_time = data.split('.')[0]
+                    df = open(filePath + '/' + province + '/' + data, encoding="GBK")
+                    for raw in df:
+                        if "城市" in raw:
+                            pass
+                        else:
+                            raw = province+','+raw.replace('\n', ' ') + ',' + date_time
+                            split = list(raw.split(','))
+                            print(split)
+                            result.append(split)
+            writer.writerows(result)
+            print(province+" finished")
+            mysql = data2mysql(result)
+            mysql[0].commit()
+            mysql[0].close()
+            mysql[1].close()
+            result.clear()
+        csvfile.close()
+        print(province + " 完成")
+# save()
